move units.Seconds to time.Seconds
