+++ xlab/base/BUILD
@@ -1,0 +1,9 @@
+load("@rules_python//python:defs.bzl", "py_library")
+
+package(default_visibility = ["//xlab:internal"])
+
+py_library(
+    name = "time",
+    srcs = ["time.py"],
+    srcs_version = "PY3",
+)
