+++ xlab/data/store/interface.py
@@ -3,8 +3,8 @@ import dataclasses
 import datetime
 from typing import Callable, List, Optional
 
+from xlab.base import time
 from xlab.data.proto import data_entry_pb2, data_type_pb2
-from xlab.data.store import units
 
 
 # Fields for retrieving data entries.
@@ -12,8 +12,8 @@ from xlab.data.store import units
 class LookupKey:
     data_space: int = 0  # Prot Enum data_entry_pb2.DataEntry.DataSpace
     symbol: Optional[str] = None
-    data_type: int = 0 # Proto Enum data_type_pb2.DataType.Enum
-    timestamp: Optional[units.Seconds] = None
+    data_type: int = 0  # Proto Enum data_type_pb2.DataType.Enum
+    timestamp: Optional[time.Seconds] = None
 
 
 class DataStore(abc.ABC):
