+++ xlab/data/store/key.py
@@ -1,7 +1,8 @@
 from typing import Tuple
 
+from xlab.base import time
 from xlab.data.proto import data_entry_pb2, data_type_pb2
-from xlab.data.store import interface, units
+from xlab.data.store import interface
 
 DataKey = Tuple[int,  # Prot Enum data_entry_pb2.DataEntry.DataSpace
                 str,  # symbol
@@ -33,5 +34,5 @@ def make_lookup_key(
     return interface.LookupKey(data_space=data_entry.data_space,
                                symbol=data_entry.symbol,
                                data_type=data_entry.data_type,
-                               timestamp=units.Seconds(
+                               timestamp=time.Seconds(
                                    data_entry.timestamp.ToSeconds()))
